+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/dto/request/CreateBarangRequestDTO.java
@@ -22,5 +22,5 @@ public class CreateBarangRequestDTO {
     private String merk;
 
     @NotNull
-    private BigInteger harga;
+    private Long harga;
 }
