+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/repository/PermintaanPengirimanDb.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface PermintaanPengirimanDb extends JpaRepository<PermintaanPengiriman, BigInteger> {
+public interface PermintaanPengirimanDb extends JpaRepository<PermintaanPengiriman, Long> {
 }
