+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/service/KaryawanService.java
@@ -10,7 +10,7 @@ public interface KaryawanService {
 
     List<Karyawan> getAllKaryawan();
 
-    Karyawan getKaryawanById(BigInteger id);
+    Karyawan getKaryawanById(Long id);
 
     long getCount();
 }
