+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/service/PermintaanPengirimanBarangService.java
@@ -10,5 +10,5 @@ public interface PermintaanPengirimanBarangService {
 
     List<PermintaanPengirimanBarang> getAllPermintaanPengirimanBarang();
 
-    PermintaanPengirimanBarang getPermintaanPengirimanBarangById(BigInteger id);
+    PermintaanPengirimanBarang getPermintaanPengirimanBarangById(Long id);
 }
