+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/service/PermintaanPengirimanService.java
@@ -10,7 +10,7 @@ public interface PermintaanPengirimanService {
 
     List<PermintaanPengiriman> getAllPermintaanPengiriman();
 
-    PermintaanPengiriman getPermintaanPengirimanById(BigInteger id);
+    PermintaanPengiriman getPermintaanPengirimanById(Long id);
 
     long getCount();
 }
